add banyak

+++ database/migrations/2021_08_02_152800_create_products_table.php
@@ -19,11 +19,11 @@ class CreateProductsTable extends Migration
             $table->foreignId('product_subcategory_id');
             $table->string('code', 20)->nullable();
             $table->string('name', 255);
-            $table->integer('weight')->nullable()->default(0);
-            $table->integer('central_stock')->nullable()->default(0);
-            $table->integer('retail_stock')->nullable()->default(0);
-            $table->integer('studio_stock')->nullable()->default(0);
-            $table->integer('bad_stock')->nullable()->default(0);
+            $table->decimal('weight')->nullable()->default(0);
+            $table->decimal('central_stock')->nullable()->default(0);
+            $table->decimal('retail_stock')->nullable()->default(0);
+            $table->decimal('studio_stock')->nullable()->default(0);
+            $table->decimal('bad_stock')->nullable()->default(0);
             $table->integer('purchase_price')->nullable()->default(0);
             $table->integer('agent_price')->nullable()->default(0);
             $table->integer('ws_price')->nullable()->default(0);
